attempting to clean up the repository
