--- scripts/jigsaw_experiment_1.sh
@@ -1,29 +1,0 @@
-# Jigsaw baseline
-
-# 100% data
-# python3 -m sss.baseline.jigsaw --fraction_data=1.0 --tpu=oai-tpu-2 --train_epochs=50
-python3 -m sss.baseline.jigsaw --fraction_data=1.0 --tpu=oai-tpu-2 --train_epochs=50
-python3 -m sss.baseline.jigsaw --fraction_data=1.0 --tpu=oai-tpu-2 --train_epochs=50
-python3 -m sss.baseline.jigsaw --fraction_data=1.0 --tpu=oai-tpu-2 --train_epochs=50
-python3 -m sss.baseline.jigsaw --fraction_data=1.0 --tpu=oai-tpu-2 --train_epochs=50
-
-# 50% data
-python3 -m sss.baseline.jigsaw --fraction_data=0.5 --tpu=oai-tpu-2 
-python3 -m sss.baseline.jigsaw --fraction_data=0.5 --tpu=oai-tpu-2 
-python3 -m sss.baseline.jigsaw --fraction_data=0.5 --tpu=oai-tpu-2 
-python3 -m sss.baseline.jigsaw --fraction_data=0.5 --tpu=oai-tpu-2 
-python3 -m sss.baseline.jigsaw --fraction_data=0.5 --tpu=oai-tpu-2 
-
-# 25% data
-python3 -m sss.baseline.jigsaw --fraction_data=0.25 --tpu=oai-tpu-2 
-python3 -m sss.baseline.jigsaw --fraction_data=0.25 --tpu=oai-tpu-2 
-python3 -m sss.baseline.jigsaw --fraction_data=0.25 --tpu=oai-tpu-2 
-python3 -m sss.baseline.jigsaw --fraction_data=0.25 --tpu=oai-tpu-2 
-python3 -m sss.baseline.jigsaw --fraction_data=0.25 --tpu=oai-tpu-2 
-
-# 10% data
-python3 -m sss.baseline.jigsaw --fraction_data=0.1 --tpu=oai-tpu-2 
-python3 -m sss.baseline.jigsaw --fraction_data=0.1 --tpu=oai-tpu-2 
-python3 -m sss.baseline.jigsaw --fraction_data=0.1 --tpu=oai-tpu-2 
-python3 -m sss.baseline.jigsaw --fraction_data=0.1 --tpu=oai-tpu-2 
-python3 -m sss.baseline.jigsaw --fraction_data=0.1 --tpu=oai-tpu-2
