--- training_logs/simclr_experiment_1/20210601-082407/train_flags.cfg
@@ -1,54 +1,0 @@
---nohelp
---nohelpfull
---nohelpshort
---nohelpxml
---noonly_check_args
---nopdb
---nopdb_post_mortem
---norun_with_pdb
---norun_with_profiling
---use_cprofile_for_profiling
---noalsologtostderr
---log_dir=
---logger_levels=
---nologtostderr
---showprefixforinfo
---stderrthreshold=fatal
---verbosity=0
---test_random_seed=301
---test_randomize_ordering_seed=
---test_srcdir=
---test_tmpdir=/tmp/absl_testing
---xml_output_file=
---runtime_oom_exit
---add_background
---backbone=resnet18
---batch_size=64
---custom_schedule
---dataset=oai
---dataset_dir=.
---eval_linear
---fraction_data=0.025
---ft_learning_rate=0.0002
---hidden_dim=512
---histdir=/home/User/Self-Supervised-Segmentation/training_logs
---image_size=288
---learning_rate=0.0002
---logdir=/home/User/Self-Supervised-Segmentation/weights
---loss_temperature=0.5
---model=simclr
---nomulti_class
---normalize
---num_cores=8
---num_head_layers=1
---online_ft
---optimizer=adam
---output_dim=128
---pretrain_encoder_only
---save_history
---nosave_weights
---tpu=oai-tpu
---train_epochs=80
---use_bfloat16
---nouse_gpu
---op_conversion_fallback_to_while_loop
